requirements.txt
streamlit
requests
beautifulsoup4
google-generativeai
tldextract